requirements.txt
# Core
pytz==2020.1
Django==3.1.3
django-configurations==2.2
gunicorn==20.0.4
newrelic==5.22.1.152
drf-yasg~=1.20.0

# For the persistence stores
psycopg2-binary==2.8.6
dj-database-url==0.5.0

# Model Tools
django-model-utils==4.0.0
django_unique_upload==0.2.1

# Rest apis
djangorestframework==3.12.2
Markdown==3.3.3
django-filter==2.4.0

# Developer Tools
ipdb==0.13.4
ipython==7.19.0
mkdocs==1.1.2
flake8==3.8.4

# Testing
mock==4.0.2
factory-boy==3.1.0
django-nose==1.4.7
nose-progressive==1.5.2
coverage==5.3

# Static and Media Storage
django-storages==1.10.1
boto3==1.16.24

